wip: clock

+++ libfp/src/utils.rs
@@ -1,8 +1,8 @@
 use embassy_time::Duration;
 use midly::num::u7;
 
-pub const fn bpm_to_clock_duration(bpm: f32, ppqn: u8) -> Duration {
-    Duration::from_nanos((1_000_000_000.0 / (bpm / 60.0 * ppqn as f32)) as u64)
+pub const fn bpm_to_clock_duration(bpm: f64, ppqn: u8) -> Duration {
+    Duration::from_nanos((1_000_000_000.0 / (bpm / 60.0 * ppqn as f64)) as u64)
 }
 
 /// Scale from 4096 to 127
